fix: Repaired the draft Sentry fix, cleared the debt-ratchet failure, and addressed the valid review
Sentry-Issue: 6529-FRONTEND-Q
