+++ utils/sentry-client-filters.ts
@@ -46,10 +46,12 @@ export {
   shouldFilterInjectedProviderProxyStartsWithError,
   shouldFilterInjectedWalletCollision,
   shouldFilterKnownWalletProviderObjectRejection,
-  shouldFilterRabbyChromeUserRejectedRequest,
-  shouldFilterRabbyMobileRainbowKitNotFoundError,
-  shouldFilterRabbyMobileUserRejectedRequest,
   shouldFilterTalismanExtensionOnboardingError,
   shouldFilterWalletConnectStaleSessionTopic,
 } from "./sentry-client-filters/wallets";
+export {
+  shouldFilterRabbyChromeUserRejectedRequest,
+  shouldFilterRabbyMobileRainbowKitNotFoundError,
+  shouldFilterRabbyMobileUserRejectedRequest,
+} from "./sentry-client-filters/rabby";
 export { __testing } from "./sentry-client-filters/testing";
