+++ utils/sentry-client-filters/wallets.ts
@@ -7,21 +7,9 @@ import {
   injectedProviderProxyStartsWithMessage,
   jsonStringifyFunction,
   metaMaskMobileUpdateUrlFunction,
-  objectCapturedPromiseRejectionMessage,
   objectCapturedPromiseRejectionWithoutStackMessage,
   providerDisconnectedCode,
   providerDisconnectedMessage,
-  rabbyChromeExtensionContentScriptUrlPrefix,
-  rabbyChromeUserRejectedCode,
-  rabbyChromeUserRejectedMessage,
-  rabbyChromeUserRejectedStackFunction,
-  rabbyChromeUserRejectedStackHeader,
-  rabbyMobileStackContextPattern,
-  rabbyMobileUserRejectedCode,
-  rabbyMobileUserRejectedMessage,
-  rabbyMobileUserRejectedStackPattern,
-  RABBY_MOBILE_RAINBOWKIT_NOT_FOUND_MESSAGE,
-  RABBY_MOBILE_USER_AGENT_TOKEN,
   readOnlyEthereumProxyBreadcrumbPattern,
   talismanExtensionOnboardingMessage,
   walletCollisionPatterns,
@@ -39,13 +27,10 @@ import type {
 import {
   getBreadcrumbMessages,
   getBreadcrumbValues,
-  getContextString,
   getEventMessage,
   getHintExceptionMessage,
   getHintExceptionStack,
   getNumericValue,
-  getRequestHeaderString,
-  getRuntimeUserAgentString,
   getSerializedExceptionStack,
   getSerializedObjectRejection,
   getStringValue,
@@ -58,8 +43,7 @@ import {
   hasAppOwnedSourceEvidence,
   hasAppOwnedSourceFrame,
   hasAppOwnedSourceStackValue,
-  hasAppOwnedStackPath,
-  hasLikelyAppOwnedFrame,
+  hasAppOwnedStackEvidence,
   hasOnlyInjectedProviderProxyFrames,
   isInjectedOrThirdPartyWalletExtensionPath,
 } from "./app-frame-utils";
@@ -76,128 +60,6 @@ import {
   hasWalletLinkWebSocketUnhandledRejectionSignature,
   isCoinbaseWalletLinkWebSocket1006Message,
 } from "./walletlink-websocket";
-
-const rabbyRainbowKitRawChunkPathPrefix = "app:///_next/static/chunks/";
-const stackFrameLocationSuffixPattern = /:\d+:\d+\)$/;
-
-function matchesStackPattern(
-  value: string | undefined,
-  pattern: string
-): boolean {
-  return value?.toLowerCase().includes(pattern) ?? false;
-}
-
-function hasRabbyMobileStackContext(
-  serializedStack: string | undefined,
-  hint?: SentryEventHint
-): boolean {
-  return [serializedStack, getHintExceptionStack(hint)].some((stack) =>
-    matchesStackPattern(stack, rabbyMobileStackContextPattern)
-  );
-}
-
-function hasRabbyMobileUserRejectedStack(
-  serializedStack: string | undefined,
-  hint?: SentryEventHint
-): boolean {
-  return [serializedStack, getHintExceptionStack(hint)].some((stack) =>
-    matchesStackPattern(stack, rabbyMobileUserRejectedStackPattern)
-  );
-}
-
-function isRabbyChromeContentScriptFrame(stackLine: string): boolean {
-  const normalizedLine = stackLine.trim();
-  return (
-    normalizedLine.startsWith("at ") &&
-    normalizedLine.includes(
-      `(${rabbyChromeExtensionContentScriptUrlPrefix}`
-    ) &&
-    stackFrameLocationSuffixPattern.test(normalizedLine)
-  );
-}
-
-function isRabbyChromeUserRejectedFrame(stackLine: string): boolean {
-  return stackLine
-    .trim()
-    .startsWith(
-      `at ${rabbyChromeUserRejectedStackFunction} (${rabbyChromeExtensionContentScriptUrlPrefix}`
-    );
-}
-
-function hasExactRabbyChromeUserRejectedStack(
-  stack: string | undefined
-): stack is string {
-  if (!stack) {
-    return false;
-  }
-
-  const [header, ...frames] = stack.split("\n");
-  return (
-    header === rabbyChromeUserRejectedStackHeader &&
-    frames.length > 0 &&
-    frames.every(isRabbyChromeContentScriptFrame) &&
-    frames.some(isRabbyChromeUserRejectedFrame)
-  );
-}
-
-function isObservedRabbyRainbowKitRawChunkFrame(
-  frame: SentryStackFrame | undefined
-): boolean {
-  if (frame?.in_app !== true || frame.function !== "n") {
-    return false;
-  }
-
-  const paths = [frame.filename, frame.abs_path].filter(
-    (path): path is string => typeof path === "string" && path.length > 0
-  );
-  return (
-    paths.length > 0 &&
-    paths.every(
-      (path) =>
-        path.startsWith(rabbyRainbowKitRawChunkPathPrefix) &&
-        path.endsWith(".js")
-    )
-  );
-}
-
-function isObservedRabbyRainbowKitNativePromiseFrame(
-  frame: SentryStackFrame | undefined
-): boolean {
-  const paths = [frame?.filename, frame?.abs_path].filter(
-    (path): path is string => typeof path === "string" && path.length > 0
-  );
-  return (
-    frame?.in_app === true &&
-    frame.function === "Promise" &&
-    paths.length > 0 &&
-    paths.every((path) => path === "[native code]")
-  );
-}
-
-function hasObservedRabbyRainbowKitRawFrames(
-  frames: SentryStackFrame[] | undefined
-): boolean {
-  // beforeSend receives this exact two-frame shape before Sentry symbolicates it.
-  return (
-    Array.isArray(frames) &&
-    frames.length === 2 &&
-    isObservedRabbyRainbowKitRawChunkFrame(frames[0]) &&
-    isObservedRabbyRainbowKitNativePromiseFrame(frames[1])
-  );
-}
-
-function hasAppOwnedStackEvidence(
-  event: SentryClientEvent,
-  serializedStack: string | undefined,
-  hint?: SentryEventHint
-): boolean {
-  const frames = event.exception?.values?.[0]?.stacktrace?.frames;
-  return (
-    hasLikelyAppOwnedFrame(frames) ||
-    hasAppOwnedStackPath(serializedStack) ||
-    hasAppOwnedStackPath(getHintExceptionStack(hint))
-  );
-}
 
 function isWalletConnectStaleSessionTopicMessage(value: string): boolean {
   const normalized = normalizeErrorPrefix(value);
@@ -418,24 +280,6 @@ function hasMetaMaskMobileUpdateUrlCircularJsonSignature(
   );
 }
 
-function hasRabbyMobileContext(event: SentryClientEvent): boolean {
-  const candidates = [
-    getContextString(event, "browser", "name"),
-    getRequestHeaderString(event, "user-agent"),
-    getRuntimeUserAgentString(),
-    getStringValue(event.tags?.["browser"]),
-    getStringValue(event.tags?.["browser.name"]),
-    getStringValue(event.tags?.["user_agent"]),
-    getStringValue(event.tags?.["userAgent"]),
-  ];
-
-  return candidates.some(
-    (candidate) =>
-      typeof candidate === "string" &&
-      candidate.toLowerCase().includes(RABBY_MOBILE_USER_AGENT_TOKEN)
-  );
-}
-
 export function shouldFilterDisconnectedWalletProviderRejection(
   event: SentryClientEvent,
   hint?: SentryEventHint
@@ -473,8 +317,7 @@ export function shouldFilterDisconnectedWalletProviderRejection(
 }
 
 function hasSingleFramelessBrowserUnhandledRejection(
-  event: SentryClientEvent,
-  expectedValue: string
+  event: SentryClientEvent
 ): boolean {
   const values = event.exception?.values;
   if (!Array.isArray(values) || values.length !== 1) {
@@ -488,7 +331,7 @@ function hasSingleFramelessBrowserUnhandledRejection(
 
   return (
     value?.type === "UnhandledRejection" &&
-    value.value === expectedValue &&
+    value.value === objectCapturedPromiseRejectionWithoutStackMessage &&
     hasBrowserUnhandledRejectionMechanism(value) &&
     hasNoFrames
   );
@@ -500,18 +343,6 @@ function hasExactCodeAndMessageShape(
   const keys = Object.keys(serialized);
   return (
     keys.length === 2 && keys.includes("code") && keys.includes("message")
-  );
-}
-
-function hasExactCodeMessageAndStackShape(
-  serialized: Record<string, unknown>
-): boolean {
-  const keys = Object.keys(serialized);
-  return (
-    keys.length === 3 &&
-    keys.includes("code") &&
-    keys.includes("message") &&
-    keys.includes("stack")
   );
 }
 
@@ -605,12 +436,7 @@ export function shouldFilterKnownWalletProviderObjectRejection(
   event: SentryClientEvent,
   hint?: SentryEventHint
 ): boolean {
-  if (
-    !hasSingleFramelessBrowserUnhandledRejection(
-      event,
-      objectCapturedPromiseRejectionWithoutStackMessage
-    )
-  ) {
+  if (!hasSingleFramelessBrowserUnhandledRejection(event)) {
     return false;
   }
 
@@ -637,105 +463,6 @@ export function shouldFilterKnownWalletProviderObjectRejection(
     message === backpackInternalJsonRpcErrorMessage &&
     hasRecentBackpackWalletCollisionBreadcrumbs(event)
   );
-}
-
-export function shouldFilterRabbyChromeUserRejectedRequest(
-  event: SentryClientEvent,
-  hint?: SentryEventHint
-): boolean {
-  if (
-    !hasSingleFramelessBrowserUnhandledRejection(
-      event,
-      objectCapturedPromiseRejectionMessage
-    )
-  ) {
-    return false;
-  }
-
-  const serialized = getSerializedObjectRejection(event, hint);
-  if (!serialized || !hasExactCodeMessageAndStackShape(serialized)) {
-    return false;
-  }
-
-  const stack = getStringValue(serialized["stack"]);
-  if (
-    serialized["code"] !== rabbyChromeUserRejectedCode ||
-    serialized["message"] !== rabbyChromeUserRejectedMessage ||
-    !hasExactRabbyChromeUserRejectedStack(stack)
-  ) {
-    return false;
-  }
-
-  return !hasAppOwnedStackEvidence(event, stack, hint);
-}
-
-export function shouldFilterRabbyMobileUserRejectedRequest(
-  event: SentryClientEvent,
-  hint?: SentryEventHint
-): boolean {
-  if (getEventMessage(event) !== objectCapturedPromiseRejectionMessage) {
-    return false;
-  }
-
-  const serialized = getSerializedObjectRejection(event, hint);
-  if (!serialized) {
-    return false;
-  }
-
-  const code = getNumericValue(serialized["code"]);
-  const message = getStringValue(serialized["message"])?.trim();
-  const stack = getStringValue(serialized["stack"]);
-
-  if (
-    code !== rabbyMobileUserRejectedCode ||
-    message !== rabbyMobileUserRejectedMessage
-  ) {
-    return false;
-  }
-
-  if (!hasRabbyMobileUserRejectedStack(stack, hint)) {
-    return false;
-  }
-
-  if (
-    !hasRabbyMobileContext(event) &&
-    !hasRabbyMobileStackContext(stack, hint)
-  ) {
-    return false;
-  }
-
-  return !hasAppOwnedStackEvidence(event, stack, hint);
-}
-
-export function shouldFilterRabbyMobileRainbowKitNotFoundError(
-  event: SentryClientEvent,
-  hint?: SentryEventHint
-): boolean {
-  const value = event.exception?.values?.[0];
-  if (
-    value?.type !== "Error" ||
-    !hasBrowserUnhandledRejectionMechanism(value)
-  ) {
-    return false;
-  }
-
-  const messageCandidates = [
-    value?.value,
-    event.message,
-    getHintExceptionMessage(hint),
-  ];
-  const hasExactMessage = messageCandidates.some(
-    (candidate) =>
-      typeof candidate === "string" &&
-      candidate.trim() === RABBY_MOBILE_RAINBOWKIT_NOT_FOUND_MESSAGE
-  );
-
-  if (!hasExactMessage) {
-    return false;
-  }
-
-  const frames = value.stacktrace?.frames;
-  return hasObservedRabbyRainbowKitRawFrames(frames);
 }
 
 export function shouldFilterCoinbaseWalletLinkWebSocket1006(
